initial set of changes, fileRouter, simplify auth middleware
